recommendation system/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.5.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
openai>=1.0.0
requests>=2.31.0
redis>=5.0.0
